requirements.txt
Django>3.2,<3.3
djangorestframework>3.12,<3.13
django-cors-headers
markdown
django-filter
drf-yasg
djangorestframework-simplejwt
pytest
pytest-django
pytest-cov
black
flake8
Faker
celery>=5.3
redis>=5.0
requests
django-redis